imp print_saved_time and failure finishes

+++ crates/cypress_parallel/src/utility.rs
@@ -95,9 +95,9 @@ pub fn collect_cy_results(results_path: &Path) -> Result<CyRunResults> {
 }
 
 /// Print how much time it would have taken in total for a single thread and how much time this app saved
-pub fn print_saved_time(total_duration: u32, time_taken: u32) {
-    let saved_time = total_duration - time_taken;
-    let saved_time_percentage = (saved_time * 100) / total_duration;
+pub fn print_saved_time(total_duration: u32, time_taken: u64) {
+    let saved_time = total_duration as u64 - time_taken;
+    let saved_time_percentage = (saved_time * 100) / total_duration as u64;
     print!(
         "Total run time: {} secs, Executed in: {} secs, Saved: {} secs ~ {}%",
         total_duration / 1000,
